+++ 1.20/Common/src/main/java/fuzs/puzzleslib/impl/core/ModContext.java
@@ -1,44 +1,55 @@
 package fuzs.puzzleslib.impl.core;
 
+import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Maps;
+import com.google.common.collect.Queues;
 import com.google.common.collect.Sets;
 import fuzs.puzzleslib.api.capability.v2.CapabilityController;
 import fuzs.puzzleslib.api.config.v3.ConfigHolder;
 import fuzs.puzzleslib.api.core.v1.Buildable;
+import fuzs.puzzleslib.api.core.v1.ContentRegistrationFlags;
+import fuzs.puzzleslib.api.core.v1.PairedModConstructor;
 import fuzs.puzzleslib.api.init.v2.RegistryManager;
 import fuzs.puzzleslib.api.network.v2.NetworkHandlerV2;
 import fuzs.puzzleslib.api.network.v3.NetworkHandlerV3;
+import net.minecraft.resources.ResourceLocation;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.Collections;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
+import java.util.*;
+import java.util.function.Predicate;
 import java.util.stream.Stream;
 
 public abstract class ModContext {
-    private static final Map<String, ModContext> CONTEXTS = Collections.synchronizedMap(Maps.newIdentityHashMap());
+    private static final Map<String, ModContext> MOD_CONTEXTS = Collections.synchronizedMap(Maps.newIdentityHashMap());
 
     final String modId;
-    private final Set<Buildable> buildables = Sets.newHashSet();
+    private final Queue<Buildable> buildables = Queues.newArrayDeque();
+    private final Map<ResourceLocation, Runnable> clientModConstructors = Maps.newHashMap();
+    private final Set<ResourceLocation> constructedPairings = Sets.newHashSet();
+    private final Set<ContentRegistrationFlags> handledFlags = EnumSet.noneOf(ContentRegistrationFlags.class);
     @Nullable NetworkHandlerV2 networkHandlerV2;
     @Nullable NetworkHandlerV3.Builder networkHandlerV3;
     @Nullable ConfigHolder.Builder configHolder;
     @Nullable RegistryManager registryManager;
     @Nullable CapabilityController capabilityController;
-    @Nullable Runnable clientModConstructor;
-    private boolean markConstructed;
 
     ModContext(String modId) {
         this.modId = modId;
     }
 
     public static ModContext get(String modId) {
-        return CONTEXTS.computeIfAbsent(modId, CommonFactories.INSTANCE::getModContext);
+        return MOD_CONTEXTS.computeIfAbsent(modId.intern(), CommonFactories.INSTANCE::getModContext);
     }
 
     public static Stream<CapabilityController> getCapabilityControllers() {
-        return CONTEXTS.values().stream().map(context -> context.capabilityController).filter(Objects::nonNull);
+        return MOD_CONTEXTS.values().stream().map(context -> context.capabilityController).filter(Objects::nonNull);
+    }
+
+    public static void testAllBuilt() {
+        for (ModContext context : MOD_CONTEXTS.values()) {
+            if (!context.buildables.isEmpty()) throw new IllegalStateException("mod context for %s has un-built buildables".formatted(context.modId));
+            if (!context.clientModConstructors.isEmpty()) throw new IllegalStateException("mod context for %s has un-built client mod constructors".formatted(context.modId));
+        }
     }
 
     public abstract NetworkHandlerV2 getNetworkHandlerV2(boolean clientAcceptsVanillaOrMissing, boolean serverAcceptsVanillaOrMissing);
@@ -52,28 +63,40 @@ public abstract class ModContext {
     public abstract CapabilityController getCapabilityController();
 
     <T extends Buildable> T addBuildable(T buildable) {
-        if (this.markConstructed) throw new IllegalStateException("mod is already constructed");
         Objects.requireNonNull(buildable, "buildable is null");
-        this.buildables.add(buildable);
+        this.buildables.offer(buildable);
         return buildable;
     }
 
-    public final void scheduleClientModConstruction(Runnable runnable) {
-        if (this.markConstructed) {
+    public final void scheduleClientModConstruction(ResourceLocation identifier, Runnable runnable) {
+        if (this.constructedPairings.contains(identifier)) {
             runnable.run();
         } else {
-            this.clientModConstructor = runnable;
+            this.clientModConstructors.put(identifier, runnable);
         }
     }
 
     public final void beforeModConstruction() {
-        this.markConstructed = true;
-        this.buildables.forEach(Buildable::build);
+        while (!this.buildables.isEmpty()) {
+            this.buildables.poll().build();
+        }
     }
 
-    public final void afterModConstruction() {
-        if (this.clientModConstructor != null) {
-            this.clientModConstructor.run();
-        }
+    public final void afterModConstruction(ResourceLocation identifier) {
+        this.constructedPairings.add(identifier);
+        Runnable runnable = this.clientModConstructors.remove(identifier);
+        if (runnable != null) runnable.run();
+    }
+
+    public final Set<ContentRegistrationFlags> getFlagsToHandle(Set<ContentRegistrationFlags> availableFlags) {
+        return availableFlags.stream()
+                .filter(Predicate.not(this.handledFlags::contains))
+                .peek(this.handledFlags::add)
+                .collect(ImmutableSet.toImmutableSet());
+    }
+
+    public static ResourceLocation getPairingIdentifier(String modId, PairedModConstructor modConstructor) {
+        ResourceLocation identifier = modConstructor.getPairingIdentifier();
+        return identifier != null ? identifier : new ResourceLocation(modId, "main");
     }
 }