+++ 1.20/Fabric/src/main/java/fuzs/puzzleslib/impl/core/FabricModConstructor.java
@@ -2,14 +2,14 @@ package fuzs.puzzleslib.impl.core;
 
 import fuzs.puzzleslib.api.core.v1.ContentRegistrationFlags;
 import fuzs.puzzleslib.api.core.v1.ModConstructor;
-import fuzs.puzzleslib.api.core.v1.context.DispenseBehaviorsContext;
 import fuzs.puzzleslib.api.item.v2.LegacySmithingTransformRecipe;
 import fuzs.puzzleslib.impl.core.context.*;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.PackType;
-import org.apache.commons.lang3.ArrayUtils;
+
+import java.util.Set;
 
 public final class FabricModConstructor {
 
@@ -17,13 +17,13 @@ public final class FabricModConstructor {
 
     }
 
-    public static void construct(ModConstructor constructor, String modId, ContentRegistrationFlags... contentRegistrations) {
-        registerContent(modId, contentRegistrations);
+    public static void construct(ModConstructor constructor, String modId, Set<ContentRegistrationFlags> availableFlags, Set<ContentRegistrationFlags> flagsToHandle) {
+        registerContent(modId, flagsToHandle);
         registerHandlers(constructor, modId);
     }
 
-    private static void registerContent(String modId, ContentRegistrationFlags[] contentRegistrations) {
-        if (ArrayUtils.contains(contentRegistrations, ContentRegistrationFlags.LEGACY_SMITHING)) {
+    private static void registerContent(String modId, Set<ContentRegistrationFlags> flagsToHandle) {
+        if (flagsToHandle.contains(ContentRegistrationFlags.LEGACY_SMITHING)) {
             Registry.register(BuiltInRegistries.RECIPE_SERIALIZER, new ResourceLocation(modId, LegacySmithingTransformRecipe.RECIPE_SERIALIZER_ID), new LegacySmithingTransformRecipe.Serializer());
         }
     }
@@ -35,7 +35,6 @@ public final class FabricModConstructor {
         constructor.onRegisterFuelBurnTimes(new FuelBurnTimesContextFabricImpl());
         constructor.onRegisterBiomeModifications(new BiomeModificationsContextFabricImpl(modId));
         constructor.onRegisterFlammableBlocks(new FlammableBlocksContextFabricImpl());
-        constructor.onRegisterDispenseBehaviors(new DispenseBehaviorsContext());
         constructor.onRegisterSpawnPlacements(new SpawnPlacementsContextFabricImpl());
         constructor.onEntityAttributeCreation(new EntityAttributesCreateContextFabricImpl());
         constructor.onEntityAttributeModification(new EntityAttributesModifyContextFabricImpl());